services/python-validator/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-genai
opencv-python
Pillow
reportlab
requests
python-multipart
pytest
pytest-asyncio
PyMuPDF
rich